% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(data, na.rm = F)
}
\arguments{
\item{data}{A data frame, should be the NOAA data}

\item{na.rm}{A logical that determines whether missing values should be omitted or not}
}
\value{
The clean data frame
}
\description{
Cleans the whole NOAA data by accomplishing all the requirments regrading lattitude, longitude,
date & location name
}
\examples{
\dontrun{
data <- readr::read_delim('signif.txt',delim ='\t')
F <- eq_clean_data(data)
T <- eq_clean_data(data,T)
}
}
