% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(.df)
}
\arguments{
\item{.df}{is the data frame inputed}
}
\value{
label the new HTML label
}
\description{
creates an HTML label of location name, eq magnitude & total deaths.
}
\examples{
\dontrun{
data <- readr::read_delim('signif.txt',delim ='\t')
data <- eq_clean_data(data,T)
eq_create_label(data)
}
}
