% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(location)
}
\arguments{
\item{location}{the location name column}
}
\value{
clean_loc : The clean location name.
}
\description{
Cleans the location name by removing the country's name with the colon using regex's.
Can be used with the function mutate() to obtain the clean column.
}
