% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leaflet.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{The input data frame}

\item{annot_col}{Column to be used as marker label}
}
\value{
map the plotted leaflet map
}
\description{
Draws an interactive map that plots the NOAA eq data.
}
\examples{
\dontrun{
data <- readr::read_delim('signif.txt',delim ='\t')
data <- eq_clean_data(data,T)
eq_map(data)
}
}
